Re-wrote gui.py so that it only contains code related to the gui
This was an attempt to stop that file becoming bloated and ugly.
All code that does something useful has been relocated to EmailBlocker.py but not all of it has been tested.
To prevent issues, the "run at startup" buttons have been temporarily disabled.

Also created a new widget: the FilterManager. This allows for easier management of email filtering rules.
It has the capacity to manage rules for sender, cc, bcc, subject, body and "fuzzy" searches. None of that fancy stuff is implemented yet, however.

TODO:
* Update EmailBlockerLite.py to work with recent changes
* Implement more complex filtering rules
* Update the CLI to allow access to these rules